This change means to solve the problem appeared in the situation when the market keeps going down, the algorithm would keep exiting and buying if RSI keeps being lower than needed and the trading volume keeps going up a lot. Now when exit strategy is triggered, RSI has to go up to a certain number first before the algorithm can buy the stock again.
Added the concept of exit strategy and re-entry strategy by files. Every
time when exit strategy is triggered, the status of stock would turn
false; When re-entry strategy is triggered, that status turns true. When
the status of a stock is false, buy strategy cannot be triggered.

Mainly modified Strategies.java to implement that logic; Add an
attribute "status" for the class Stock; Modified Main.java to read in
exitStrategies.txt and reentryStrategies.txt. Now exit strategies are
not put in sellStrategies.txt.