Design Modification: 'Reentry' and 'Exit' as Operators
Implemented 'Reentry' and 'Exit' as operator signals (like NOT and AND).
They both in a signal object and switche stock.status if the signal is
true;
The signals isReentryByRsiHigherThanSignal and
isExitByClosingDownToSignal have been deleted. The corresponding test
cases have been modified to adopt the changes;
'Interperter' is changed to 'Interpreter'